modul-10/requirements.txt
requests
openai
langchain
langchain-openai
langchain-community
langchain-text-splitters
azure-core
azure-identity
azure-search-documents
pypdf
python-dotenv
pandas
funcy
ragas==0.1.2